+++ src/models/InventoryLocation.php
@@ -16,6 +16,7 @@ use craft\commerce\Plugin;
 use craft\commerce\records\InventoryLocation as InventoryLocationRecord;
 use craft\elements\Address;
 use craft\helpers\UrlHelper;
+use craft\validators\HandleValidator;
 use craft\validators\UniqueValidator;
 use DateTime;
 
@@ -163,6 +164,12 @@ class InventoryLocation extends Model implements Chippable, CpEditable, Actionab
             'targetClass' => InventoryLocationRecord::class,
             'targetAttribute' => 'handle',
             'message' => Craft::t('yii', '{attribute} "{value}" has already been taken.'),
+        ];
+
+        $rules[] = [
+            ['handle'],
+            HandleValidator::class,
+            'reservedWords' => ['id', 'dateCreated', 'dateUpdated', 'uid', 'title', 'create'],
         ];
 
         return $rules;